backend/utils/requirements.txt
flask
torch
torchvision
torchaudio
pillow
requests
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
